pyproject.toml
[project]
name = "ingenieriaweb-grupo11"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.5",
    "django-cloudinary-storage>=0.3.0",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "whitenoise[brotli]>=6.9.0",
    "python-dotenv>=1.1.1",
    "django-haystack>=3.3.0",
    "whoosh>=2.7.4",
]
